requirements-local.txt
# --- CORE AI & AUDIO ---
faster-whisper==0.10.0
ollama
transformers
ctranslate2
pyaudio
numpy
RealtimeTTS
pyttsx3
keyboard

# --- BACKEND & SECURITY ---
flask
flask-cors
python-dotenv
cryptography

# --- SYSTEM & AUTOMATION ---
psutil
pyautogui
opencv-python
screen-brightness-control
pyperclip

# --- WEB & SCRAPING ---
requests
urllib3
beautifulsoup4
lxml
wikipedia
googlesearch-python

# --- UTILITIES & ELITE SKILLS ---
yfinance             # For Market Analyst
PyPDF2               # For Document Intel (PDFs)
python-docx          # For Document Intel (Word)
speedtest-cli        # For Network Monitor
pywhatkit            # For WhatsApp
ffmpeg-python        # For Media processing
thefuzz[speedup]     # For NLP matching
python-Levenshtein

google-generativeai